refactor(ui): split ExploitBlockLogs into table/filters/detail-modal
