+++ ui/src/components/exploit-block-logs/ExploitLogDetailModal.tsx
@@ -1,0 +1,418 @@
+import { useState, useEffect } from 'react';
+import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
+import { useTranslation } from 'react-i18next';
+import { addGlobalExploitExclusion, addHostExploitExclusion, fetchExploitRuleHosts } from '../../api/exploit-rules';
+import { banIP } from '../../api/security';
+import { api } from '../../api/client';
+import type { Log } from '../../types/log';
+import { useEscapeKey } from '../../hooks/useEscapeKey';
+import { StatusCodeBadge, ExploitRuleBadge } from './ExploitLogTable';
+
+interface ExploitLogDetailModalProps {
+  log: Log;
+  ruleName?: string;
+  onClose: () => void;
+  onExceptionAdded?: () => void;
+}
+
+export function ExploitLogDetailModal({ log, ruleName, onClose, onExceptionAdded }: ExploitLogDetailModalProps) {
+  const { t, i18n } = useTranslation('exploitLogs');
+  const queryClient = useQueryClient();
+  const [showExceptionForm, setShowExceptionForm] = useState(false);
+  const [exceptionType, setExceptionType] = useState<'global' | 'host'>('global');
+  const [selectedHostId, setSelectedHostId] = useState<string>('');
+  const [exceptionReason, setExceptionReason] = useState('');
+  const [exceptionMessage, setExceptionMessage] = useState<{ type: 'success' | 'error'; text: string } | null>(null);
+
+  // IP Ban states
+  const [showBanForm, setShowBanForm] = useState(false);
+  const [banReason, setBanReason] = useState('');
+  const [banDuration, setBanDuration] = useState<number | undefined>(undefined);
+  const [isGlobalBan, setIsGlobalBan] = useState(false);
+  const [proxyHostId, setProxyHostId] = useState<string | null>(null);
+
+  useEscapeKey(onClose);
+
+  // Fetch proxy host ID for the current host (for per-host bans)
+  const { data: hostData } = useQuery({
+    queryKey: ['proxyHostByDomain', log.host],
+    queryFn: async () => {
+      if (!log.host) return null;
+      const response = await api.get<{ id: string }>(`/api/v1/proxy-hosts/by-domain/${encodeURIComponent(log.host)}`);
+      return response.id || null;
+    },
+    enabled: !!log.host && showBanForm,
+    staleTime: 60000,
+  });
+
+  // Update proxyHostId when data is fetched
+  useEffect(() => {
+    if (hostData) {
+      setProxyHostId(hostData);
+    }
+  }, [hostData]);
+
+  // Fetch hosts for host-specific exceptions
+  const { data: hostsData } = useQuery({
+    queryKey: ['exploit-rule-hosts'],
+    queryFn: fetchExploitRuleHosts,
+    enabled: showExceptionForm,
+  });
+
+  const globalExclusionMutation = useMutation({
+    mutationFn: () => addGlobalExploitExclusion(log.exploit_rule!, exceptionReason || `Excluded from log: ${log.request_uri}`),
+    onSuccess: () => {
+      setExceptionMessage({ type: 'success', text: t('modal.exception.globalSuccess') });
+      queryClient.invalidateQueries({ queryKey: ['exploit-rules'] });
+      setShowExceptionForm(false);
+      onExceptionAdded?.();
+    },
+    onError: (error: Error) => {
+      setExceptionMessage({ type: 'error', text: error.message });
+    },
+  });
+
+  const hostExclusionMutation = useMutation({
+    mutationFn: () => addHostExploitExclusion(selectedHostId, log.exploit_rule!, exceptionReason || `Excluded from log: ${log.request_uri}`),
+    onSuccess: () => {
+      setExceptionMessage({ type: 'success', text: t('modal.exception.hostSuccess') });
+      queryClient.invalidateQueries({ queryKey: ['exploit-rules'] });
+      setShowExceptionForm(false);
+      onExceptionAdded?.();
+    },
+    onError: (error: Error) => {
+      setExceptionMessage({ type: 'error', text: error.message });
+    },
+  });
+
+  // IP Ban mutation
+  const banMutation = useMutation({
+    mutationFn: () => {
+      if (!log.client_ip) {
+        throw new Error('Missing client IP');
+      }
+      const banData: { ip_address: string; reason?: string; ban_time?: number; proxy_host_id?: string } = {
+        ip_address: log.client_ip,
+        reason: banReason || `Blocked from exploit log - ${log.exploit_rule || 'N/A'}: ${log.request_uri || 'N/A'}`,
+        ban_time: banDuration,
+      };
+      if (!isGlobalBan && proxyHostId) {
+        banData.proxy_host_id = proxyHostId;
+      }
+      return banIP(banData);
+    },
+    onSuccess: () => {
+      queryClient.invalidateQueries({ queryKey: ['banned-ips'] });
+      alert(isGlobalBan
+        ? t('banIP.successGlobal', { ip: log.client_ip, defaultValue: '{{ip}}가 전역 차단되었습니다.' })
+        : t('banIP.success', { ip: log.client_ip, defaultValue: '{{ip}}가 차단되었습니다.' }));
+      setShowBanForm(false);
+      setBanReason('');
+      setBanDuration(undefined);
+      setIsGlobalBan(false);
+    },
+    onError: (error) => {
+      alert(t('banIP.failed', { error: (error as Error).message, defaultValue: 'IP 차단 실패: {{error}}' }));
+    },
+  });
+
+  const handleAddException = () => {
+    if (exceptionType === 'global') {
+      globalExclusionMutation.mutate();
+    } else if (selectedHostId) {
+      hostExclusionMutation.mutate();
+    }
+  };
+
+  return (
+    <div className="fixed inset-0 bg-black/50 flex items-center justify-center z-50 p-4" onClick={onClose}>
+      <div className="bg-white dark:bg-slate-800 rounded-xl shadow-2xl max-w-3xl w-full max-h-[90vh] overflow-hidden" onClick={e => e.stopPropagation()}>
+        <div className="px-6 py-4 border-b border-slate-200 dark:border-slate-700 flex items-center justify-between bg-red-50 dark:bg-red-900/20">
+          <h3 className="text-lg font-semibold text-red-900 dark:text-red-100">{t('modal.title')}</h3>
+          <button onClick={onClose} className="text-slate-400 hover:text-slate-600 dark:hover:text-slate-300">
+            <svg className="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
+            </svg>
+          </button>
+        </div>
+        <div className="p-6 overflow-y-auto max-h-[calc(90vh-80px)] space-y-4">
+          <div className="grid grid-cols-2 gap-4">
+            <div>
+              <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.timestamp')}</label>
+              <p className="text-sm text-slate-900 dark:text-white">{new Date(log.timestamp).toLocaleString(i18n.language)}</p>
+            </div>
+            <div>
+              <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.exploitRule')}</label>
+              <p className="mt-1">
+                <ExploitRuleBadge ruleId={log.exploit_rule} ruleName={ruleName} />
+              </p>
+            </div>
+            <div>
+              <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.clientIp')}</label>
+              <div className="flex items-center gap-2">
+                {log.client_ip ? (
+                  <button
+                    onClick={() => setShowBanForm(!showBanForm)}
+                    className="group flex items-center gap-2 px-2 py-1 -ml-2 text-sm font-mono text-slate-900 dark:text-white hover:bg-red-50 dark:hover:bg-red-900/30 hover:text-red-700 dark:hover:text-red-400 rounded-lg transition-colors"
+                    title={t('banIP.clickToBan', { defaultValue: 'IP 차단하려면 클릭' })}
+                  >
+                    <span className={log.is_banned ? 'text-red-600 dark:text-red-400' : ''}>{log.client_ip}</span>
+                    {log.is_banned && (
+                      <span className="px-1.5 py-0.5 bg-red-100 dark:bg-red-900/30 text-red-700 dark:text-red-400 text-xs font-medium rounded">
+                        Banned
+                      </span>
+                    )}
+                    <svg className="w-4 h-4 text-red-500 opacity-0 group-hover:opacity-100 transition-opacity" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M18.364 18.364A9 9 0 005.636 5.636m12.728 12.728A9 9 0 015.636 5.636m12.728 12.728L5.636 5.636" />
+                    </svg>
+                  </button>
+                ) : (
+                  <p className="text-sm text-slate-900 dark:text-white font-mono">-</p>
+                )}
+              </div>
+            </div>
+            <div>
+              <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.statusCode')}</label>
+              <p className="mt-1">{log.status_code && <StatusCodeBadge code={log.status_code} />}</p>
+            </div>
+            <div>
+              <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.host')}</label>
+              <p className="text-sm text-slate-900 dark:text-white">{log.host || '-'}</p>
+            </div>
+            <div>
+              <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.method')}</label>
+              <p className="text-sm text-slate-900 dark:text-white font-mono">{log.request_method || '-'}</p>
+            </div>
+            <div>
+              <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.country')}</label>
+              <p className="text-sm text-slate-900 dark:text-white">{log.geo_country_code || '-'}</p>
+            </div>
+          </div>
+
+          {/* IP Ban Form */}
+          {showBanForm && log.client_ip && (
+            <div className="p-4 bg-red-50 dark:bg-red-900/20 border border-red-200 dark:border-red-800 rounded-lg">
+              <h4 className="text-sm font-semibold text-red-800 dark:text-red-300 mb-3">{t('banIP.title', { ip: log.client_ip, defaultValue: '{{ip}} 차단' })}</h4>
+              <div className="space-y-3">
+                {/* Scope Toggle */}
+                <div>
+                  <label className="text-xs font-medium text-red-700 dark:text-red-400 uppercase mb-2 block">
+                    {t('banIP.scope', { defaultValue: '적용 범위' })}
+                  </label>
+                  <div className="flex gap-2">
+                    <button
+                      type="button"
+                      onClick={() => setIsGlobalBan(false)}
+                      className={`flex-1 px-3 py-2 text-sm font-medium rounded-lg transition-colors ${
+                        !isGlobalBan
+                          ? 'bg-red-600 text-white'
+                          : 'bg-white dark:bg-slate-800 text-red-700 dark:text-red-400 border border-red-300 dark:border-red-700 hover:bg-red-50 dark:hover:bg-red-900/30'
+                      }`}
+                    >
+                      {t('banIP.perHost', { defaultValue: '이 호스트만', host: log.host })}
+                    </button>
+                    <button
+                      type="button"
+                      onClick={() => setIsGlobalBan(true)}
+                      className={`flex-1 px-3 py-2 text-sm font-medium rounded-lg transition-colors ${
+                        isGlobalBan
+                          ? 'bg-purple-600 text-white'
+                          : 'bg-white dark:bg-slate-800 text-purple-700 dark:text-purple-400 border border-purple-300 dark:border-purple-700 hover:bg-purple-50 dark:hover:bg-purple-900/30'
+                      }`}
+                    >
+                      {t('banIP.global', { defaultValue: '전역 차단 (모든 호스트)' })}
+                    </button>
+                  </div>
+                  {isGlobalBan && (
+                    <p className="mt-1 text-xs text-purple-600 dark:text-purple-400">
+                      {t('banIP.globalDescription', { defaultValue: '이 IP는 모든 프록시 호스트에서 차단됩니다.' })}
+                    </p>
+                  )}
+                </div>
+                <div>
+                  <label className="text-xs font-medium text-red-700 dark:text-red-400 uppercase mb-1 block">
+                    {t('banIP.reason', { defaultValue: '차단 사유' })}
+                  </label>
+                  <input
+                    type="text"
+                    value={banReason}
+                    onChange={(e) => setBanReason(e.target.value)}
+                    placeholder={t('banIP.reasonPlaceholder', { defaultValue: '차단 사유를 입력하세요 (선택)' })}
+                    className="w-full px-3 py-2 border border-red-300 dark:border-red-700 rounded-lg text-sm focus:ring-2 focus:ring-red-500 focus:border-red-500 bg-white dark:bg-slate-800 text-slate-900 dark:text-white"
+                  />
+                </div>
+                <div>
+                  <label className="text-xs font-medium text-red-700 dark:text-red-400 uppercase mb-1 block">
+                    {t('banIP.duration', { defaultValue: '차단 기간' })}
+                  </label>
+                  <select
+                    value={banDuration ?? ''}
+                    onChange={(e) => setBanDuration(e.target.value ? parseInt(e.target.value) : undefined)}
+                    className="w-full px-3 py-2 border border-red-300 dark:border-red-700 rounded-lg text-sm focus:ring-2 focus:ring-red-500 focus:border-red-500 bg-white dark:bg-slate-800 text-slate-900 dark:text-white"
+                  >
+                    <option value="">{t('banIP.permanent', { defaultValue: '영구 차단' })}</option>
+                    <option value="3600">{t('banIP.duration1h', { defaultValue: '1시간' })}</option>
+                    <option value="86400">{t('banIP.duration1d', { defaultValue: '1일' })}</option>
+                    <option value="604800">{t('banIP.duration1w', { defaultValue: '1주일' })}</option>
+                    <option value="2592000">{t('banIP.duration1m', { defaultValue: '1개월' })}</option>
+                  </select>
+                </div>
+                <div className="flex gap-2">
+                  <button
+                    onClick={() => banMutation.mutate()}
+                    disabled={banMutation.isPending}
+                    className={`flex-1 px-4 py-2 text-sm font-medium text-white rounded-lg transition-colors disabled:opacity-50 ${
+                      isGlobalBan ? 'bg-purple-600 hover:bg-purple-700' : 'bg-red-600 hover:bg-red-700'
+                    }`}
+                  >
+                    {banMutation.isPending ? t('banIP.processing', { defaultValue: '처리 중...' }) : (isGlobalBan ? t('banIP.submitGlobal', { defaultValue: '전역 차단' }) : t('banIP.submit', { defaultValue: '차단' }))}
+                  </button>
+                  <button
+                    onClick={() => {
+                      setShowBanForm(false);
+                      setBanReason('');
+                      setBanDuration(undefined);
+                      setIsGlobalBan(false);
+                    }}
+                    className="px-4 py-2 text-sm font-medium text-slate-700 dark:text-slate-300 bg-slate-100 dark:bg-slate-700 hover:bg-slate-200 dark:hover:bg-slate-600 rounded-lg transition-colors"
+                  >
+                    {t('banIP.cancel', { defaultValue: '취소' })}
+                  </button>
+                </div>
+              </div>
+            </div>
+          )}
+
+          {ruleName && (
+            <div>
+              <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.ruleName')}</label>
+              <p className="text-sm text-slate-900 dark:text-white bg-red-50 dark:bg-red-900/30 p-2 rounded">{ruleName}</p>
+            </div>
+          )}
+
+          <div>
+            <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.requestUri')}</label>
+            <p className="text-sm text-slate-900 dark:text-white font-mono break-all bg-slate-50 dark:bg-slate-700 p-2 rounded">{log.request_uri || '-'}</p>
+          </div>
+
+          <div>
+            <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.userAgent')}</label>
+            <p className="text-sm text-slate-900 dark:text-white font-mono break-all bg-slate-50 dark:bg-slate-700 p-2 rounded">{log.http_user_agent || '-'}</p>
+          </div>
+
+          {log.raw_log && (
+            <div>
+              <label className="text-xs text-slate-500 dark:text-slate-400 font-medium">{t('modal.rawLog')}</label>
+              <pre className="text-xs text-slate-600 dark:text-slate-300 bg-slate-100 dark:bg-slate-700 p-3 rounded overflow-x-auto whitespace-pre-wrap break-all">{log.raw_log}</pre>
+            </div>
+          )}
+
+          {/* Exception Message */}
+          {exceptionMessage && (
+            <div className={`px-4 py-3 rounded-lg text-sm font-medium ${exceptionMessage.type === 'success'
+              ? 'bg-green-50 dark:bg-green-900/30 text-green-700 dark:text-green-400 border border-green-200 dark:border-green-800'
+              : 'bg-red-50 dark:bg-red-900/30 text-red-700 dark:text-red-400 border border-red-200 dark:border-red-800'
+              }`}>
+              {exceptionMessage.text}
+            </div>
+          )}
+
+          {/* Exception Form */}
+          {log.exploit_rule && (
+            <div className="border-t border-slate-200 dark:border-slate-700 pt-4 mt-4">
+              {!showExceptionForm ? (
+                <button
+                  onClick={() => setShowExceptionForm(true)}
+                  className="w-full px-4 py-2 bg-amber-100 dark:bg-amber-900/30 text-amber-800 dark:text-amber-200 rounded-lg hover:bg-amber-200 dark:hover:bg-amber-900/50 transition-colors flex items-center justify-center gap-2"
+                >
+                  <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 9v3m0 0v3m0-3h3m-3 0H9m12 0a9 9 0 11-18 0 9 9 0 0118 0z" />
+                  </svg>
+                  {t('modal.exception.addButton')}
+                </button>
+              ) : (
+                <div className="bg-amber-50 dark:bg-amber-900/20 border border-amber-200 dark:border-amber-800 rounded-lg p-4 space-y-3">
+                  <h4 className="text-sm font-semibold text-amber-900 dark:text-amber-100">{t('modal.exception.title')}</h4>
+
+                  {/* Exception Type Selection */}
+                  <div className="flex gap-4">
+                    <label className="flex items-center gap-2 cursor-pointer">
+                      <input
+                        type="radio"
+                        name="exceptionType"
+                        value="global"
+                        checked={exceptionType === 'global'}
+                        onChange={() => setExceptionType('global')}
+                        className="w-4 h-4 text-amber-600"
+                      />
+                      <span className="text-sm text-slate-700 dark:text-slate-300">{t('modal.exception.global')}</span>
+                    </label>
+                    <label className="flex items-center gap-2 cursor-pointer">
+                      <input
+                        type="radio"
+                        name="exceptionType"
+                        value="host"
+                        checked={exceptionType === 'host'}
+                        onChange={() => setExceptionType('host')}
+                        className="w-4 h-4 text-amber-600"
+                      />
+                      <span className="text-sm text-slate-700 dark:text-slate-300">{t('modal.exception.perHost')}</span>
+                    </label>
+                  </div>
+
+                  {/* Host Selection (only for per-host exception) */}
+                  {exceptionType === 'host' && (
+                    <div>
+                      <label className="block text-xs text-slate-500 dark:text-slate-400 mb-1">{t('modal.exception.selectHost')}</label>
+                      <select
+                        value={selectedHostId}
+                        onChange={(e) => setSelectedHostId(e.target.value)}
+                        className="w-full px-3 py-2 border border-slate-300 dark:border-slate-600 rounded-lg text-sm bg-white dark:bg-slate-700 text-slate-900 dark:text-white"
+                      >
+                        <option value="">{t('modal.exception.selectHostPlaceholder')}</option>
+                        {hostsData?.hosts?.map((host) => (
+                          <option key={host.proxy_host_id} value={host.proxy_host_id}>
+                            {host.proxy_host_name}
+                          </option>
+                        ))}
+                      </select>
+                    </div>
+                  )}
+
+                  {/* Reason Input */}
+                  <div>
+                    <label className="block text-xs text-slate-500 dark:text-slate-400 mb-1">{t('modal.exception.reason')}</label>
+                    <input
+                      type="text"
+                      value={exceptionReason}
+                      onChange={(e) => setExceptionReason(e.target.value)}
+                      placeholder={t('modal.exception.reasonPlaceholder')}
+                      className="w-full px-3 py-2 border border-slate-300 dark:border-slate-600 rounded-lg text-sm bg-white dark:bg-slate-700 text-slate-900 dark:text-white"
+                    />
+                  </div>
+
+                  {/* Actions */}
+                  <div className="flex gap-2 justify-end">
+                    <button
+                      onClick={() => setShowExceptionForm(false)}
+                      className="px-4 py-2 text-sm text-slate-600 dark:text-slate-400 hover:text-slate-800 dark:hover:text-slate-200"
+                    >
+                      {t('modal.exception.cancel')}
+                    </button>
+                    <button
+                      onClick={handleAddException}
+                      disabled={exceptionType === 'host' && !selectedHostId || globalExclusionMutation.isPending || hostExclusionMutation.isPending}
+                      className="px-4 py-2 bg-amber-600 text-white rounded-lg hover:bg-amber-700 transition-colors text-sm font-medium disabled:opacity-50 disabled:cursor-not-allowed"
+                    >
+                      {globalExclusionMutation.isPending || hostExclusionMutation.isPending ? t('modal.exception.adding') : t('modal.exception.confirm')}
+                    </button>
+                  </div>
+                </div>
+              )}
+            </div>
+          )}
+        </div>
+      </div>
+    </div>
+  );
+}
