+++ ui/src/components/exploit-block-logs/ExploitLogFilters.tsx
@@ -1,0 +1,137 @@
+import { useTranslation } from 'react-i18next';
+import DatePicker from 'react-datepicker';
+import type { Locale } from 'date-fns';
+import { HelpTip } from '../common/HelpTip';
+
+interface ExploitLogFiltersProps {
+  startDate: Date | null;
+  endDate: Date | null;
+  setStartDate: (d: Date | null) => void;
+  setEndDate: (d: Date | null) => void;
+  dateLocale: Locale;
+
+  ruleFilter: string;
+  setRuleFilter: (v: string) => void;
+  ruleOptions: { id: string; name: string }[];
+
+  searchInput: string;
+  setSearchInput: (v: string) => void;
+
+  perPage: number;
+  setPerPage: (n: number) => void;
+
+  onPageReset: () => void;
+}
+
+export function ExploitLogFilters({
+  startDate,
+  endDate,
+  setStartDate,
+  setEndDate,
+  dateLocale,
+  ruleFilter,
+  setRuleFilter,
+  ruleOptions,
+  searchInput,
+  setSearchInput,
+  perPage,
+  setPerPage,
+  onPageReset,
+}: ExploitLogFiltersProps) {
+  const { t } = useTranslation('exploitLogs');
+
+  return (
+    <div className="bg-white dark:bg-slate-800 rounded-lg border border-slate-200 dark:border-slate-700 p-4">
+      <div className="flex flex-wrap gap-4 items-end">
+        {/* Date Range */}
+        <div className="flex gap-2 items-center">
+          <div>
+            <label className="block text-xs text-slate-500 dark:text-slate-400 mb-1 flex items-center gap-1">
+              {t('filters.start')}
+              <HelpTip content={t('filters.startHelp')} />
+            </label>
+            <DatePicker
+              selected={startDate}
+              onChange={setStartDate}
+              showTimeSelect
+              timeFormat="HH:mm"
+              timeIntervals={30}
+              dateFormat="MM/dd HH:mm"
+              locale={dateLocale}
+              className="px-3 py-1.5 border border-slate-300 dark:border-slate-600 rounded-lg text-sm w-36 bg-white dark:bg-slate-700 text-slate-900 dark:text-white"
+            />
+          </div>
+          <span className="text-slate-400 mt-5">~</span>
+          <div>
+            <label className="block text-xs text-slate-500 dark:text-slate-400 mb-1 flex items-center gap-1">
+              {t('filters.end')}
+              <HelpTip content={t('filters.endHelp')} />
+            </label>
+            <DatePicker
+              selected={endDate}
+              onChange={setEndDate}
+              showTimeSelect
+              timeFormat="HH:mm"
+              timeIntervals={30}
+              dateFormat="MM/dd HH:mm"
+              locale={dateLocale}
+              className="px-3 py-1.5 border border-slate-300 dark:border-slate-600 rounded-lg text-sm w-36 bg-white dark:bg-slate-700 text-slate-900 dark:text-white"
+            />
+          </div>
+        </div>
+
+        {/* Rule Filter */}
+        <div>
+          <label className="block text-xs text-slate-500 dark:text-slate-400 mb-1 flex items-center gap-1">
+            {t('filters.rule')}
+            <HelpTip content={t('filters.ruleHelp')} />
+          </label>
+          <select
+            value={ruleFilter}
+            onChange={e => { setRuleFilter(e.target.value); onPageReset(); }}
+            className="px-3 py-1.5 border border-slate-300 dark:border-slate-600 rounded-lg text-sm bg-white dark:bg-slate-700 text-slate-900 dark:text-white min-w-[200px]"
+          >
+            <option value="">{t('filters.allRules')}</option>
+            {ruleOptions.map(rule => (
+              <option key={rule.id} value={rule.id}>
+                {rule.name}
+              </option>
+            ))}
+          </select>
+        </div>
+
+        {/* Search */}
+        <div className="flex-1 min-w-[200px]">
+          <label className="block text-xs text-slate-500 dark:text-slate-400 mb-1 flex items-center gap-1">
+            {t('filters.search')}
+            <HelpTip content={t('filters.searchHelp')} />
+          </label>
+          <input
+            type="text"
+            value={searchInput}
+            onChange={e => { setSearchInput(e.target.value); onPageReset(); }}
+            placeholder={t('filters.searchPlaceholder')}
+            className="w-full px-3 py-1.5 border border-slate-300 dark:border-slate-600 rounded-lg text-sm bg-white dark:bg-slate-700 text-slate-900 dark:text-white"
+          />
+        </div>
+
+        {/* Per Page */}
+        <div>
+          <label className="block text-xs text-slate-500 dark:text-slate-400 mb-1 flex items-center gap-1">
+            {t('filters.perPage')}
+            <HelpTip content={t('filters.perPageHelp')} />
+          </label>
+          <select
+            value={perPage}
+            onChange={e => { setPerPage(Number(e.target.value)); onPageReset(); }}
+            className="px-3 py-1.5 border border-slate-300 dark:border-slate-600 rounded-lg text-sm bg-white dark:bg-slate-700 text-slate-900 dark:text-white"
+          >
+            <option value={25}>25</option>
+            <option value={50}>50</option>
+            <option value={100}>100</option>
+          </select>
+        </div>
+      </div>
+    </div>
+  );
+}
