+++ ui/src/components/exploit-block-logs/ExploitLogTable.tsx
@@ -1,0 +1,175 @@
+import { useTranslation } from 'react-i18next';
+import type { Log } from '../../types/log';
+
+export function formatTime(timestamp: string, language: string = 'ko'): string {
+  return new Date(timestamp).toLocaleString(language);
+}
+
+export function StatusCodeBadge({ code }: { code: number }) {
+  let color = 'bg-gray-100 text-gray-800 dark:bg-gray-800 dark:text-gray-300';
+  if (code >= 200 && code < 300) color = 'bg-green-100 text-green-800 dark:bg-green-900/30 dark:text-green-400';
+  else if (code >= 300 && code < 400) color = 'bg-blue-100 text-blue-800 dark:bg-blue-900/30 dark:text-blue-400';
+  else if (code >= 400 && code < 500) color = 'bg-yellow-100 text-yellow-800 dark:bg-yellow-900/30 dark:text-yellow-400';
+  else if (code >= 500) color = 'bg-red-100 text-red-800 dark:bg-red-900/30 dark:text-red-400';
+  return (
+    <span className={`px-2 py-0.5 rounded text-xs font-medium ${color}`}>
+      {code}
+    </span>
+  );
+}
+
+export function ExploitRuleBadge({ ruleId, ruleName }: { ruleId?: string; ruleName?: string }) {
+  if (!ruleId) return <span className="text-slate-400">-</span>;
+
+  // Determine color and icon based on rule name or category
+  let color = 'bg-red-100 dark:bg-red-900/30 text-red-800 dark:text-red-300';
+  let icon = '🛡️';
+  const displayName = ruleName || ruleId;
+  const lowerName = displayName.toLowerCase();
+
+  if (lowerName.includes('sql') || lowerName.includes('injection')) {
+    color = 'bg-orange-100 dark:bg-orange-900/30 text-orange-800 dark:text-orange-300';
+    icon = '💉';
+  } else if (lowerName.includes('rfi') || lowerName.includes('lfi') || lowerName.includes('file') || lowerName.includes('traversal')) {
+    color = 'bg-purple-100 dark:bg-purple-900/30 text-purple-800 dark:text-purple-300';
+    icon = '📂';
+  } else if (lowerName.includes('scan') || lowerName.includes('probe')) {
+    color = 'bg-yellow-100 dark:bg-yellow-900/30 text-yellow-800 dark:text-yellow-300';
+    icon = '🔍';
+  } else if (lowerName.includes('method') || lowerName.includes('http')) {
+    color = 'bg-blue-100 dark:bg-blue-900/30 text-blue-800 dark:text-blue-300';
+    icon = '📡';
+  } else if (lowerName.includes('vcs') || lowerName.includes('git') || lowerName.includes('.env')) {
+    color = 'bg-red-100 dark:bg-red-900/30 text-red-800 dark:text-red-300';
+    icon = '📁';
+  } else if (lowerName.includes('xss') || lowerName.includes('script')) {
+    color = 'bg-pink-100 dark:bg-pink-900/30 text-pink-800 dark:text-pink-300';
+    icon = '⚠️';
+  }
+
+  // Show rule name if available, otherwise truncated ID
+  const label = ruleName || `${ruleId.substring(0, 8)}...`;
+
+  return (
+    <span className={`inline-flex items-center gap-1 px-2 py-0.5 rounded text-xs font-medium ${color}`} title={displayName}>
+      <span>{icon}</span>
+      <span className="max-w-[150px] truncate">{label}</span>
+    </span>
+  );
+}
+
+interface ExploitLogTableProps {
+  data?: {
+    data?: Log[];
+    total?: number;
+    total_pages?: number;
+    page?: number;
+    has_more?: boolean;
+  };
+  isLoading: boolean;
+  error: unknown;
+  ruleNameMap: Record<string, string>;
+  onRowClick: (log: Log) => void;
+  page: number;
+  onPageChange: (p: number | ((p: number) => number)) => void;
+}
+
+export function ExploitLogTable({ data, isLoading, error, ruleNameMap, onRowClick, page, onPageChange }: ExploitLogTableProps) {
+  const { t, i18n } = useTranslation('exploitLogs');
+
+  return (
+    <div className="bg-white dark:bg-slate-800 rounded-lg border border-slate-200 dark:border-slate-700 overflow-hidden">
+      {isLoading ? (
+        <div className="p-8 text-center text-slate-500 dark:text-slate-400">{t('loading')}</div>
+      ) : error ? (
+        <div className="p-8 text-center text-red-500">{t('error')}</div>
+      ) : !data?.data?.length ? (
+        <div className="p-8 text-center text-slate-500 dark:text-slate-400">
+          <div className="text-4xl mb-2">🎉</div>
+          <div>{t('empty')}</div>
+        </div>
+      ) : (
+        <div className="overflow-x-auto">
+          <table className="w-full">
+            <thead className="bg-red-50 dark:bg-red-900/20 border-b border-red-100 dark:border-red-800">
+              <tr>
+                <th className="px-4 py-3 text-left text-xs font-semibold text-red-700 dark:text-red-300">{t('table.time')}</th>
+                <th className="px-4 py-3 text-left text-xs font-semibold text-red-700 dark:text-red-300">{t('table.rule')}</th>
+                <th className="px-4 py-3 text-left text-xs font-semibold text-red-700 dark:text-red-300">{t('table.clientIp')}</th>
+                <th className="px-4 py-3 text-left text-xs font-semibold text-red-700 dark:text-red-300">{t('table.country')}</th>
+                <th className="px-4 py-3 text-left text-xs font-semibold text-red-700 dark:text-red-300">{t('table.host')}</th>
+                <th className="px-4 py-3 text-left text-xs font-semibold text-red-700 dark:text-red-300">{t('table.uri')}</th>
+                <th className="px-4 py-3 text-left text-xs font-semibold text-red-700 dark:text-red-300">{t('table.status')}</th>
+              </tr>
+            </thead>
+            <tbody className="divide-y divide-slate-100 dark:divide-slate-700">
+              {data.data.map((log) => (
+                <tr
+                  key={log.id}
+                  className="hover:bg-red-50 dark:hover:bg-slate-700/50 cursor-pointer transition-colors"
+                  onClick={() => onRowClick(log)}
+                >
+                  <td className="px-4 py-3 text-xs text-slate-500 dark:text-slate-400 whitespace-nowrap">
+                    {formatTime(log.timestamp, i18n.language)}
+                  </td>
+                  <td className="px-4 py-3 whitespace-nowrap">
+                    <ExploitRuleBadge ruleId={log.exploit_rule} ruleName={ruleNameMap[log.exploit_rule || '']} />
+                  </td>
+                  <td className="px-4 py-3 whitespace-nowrap">
+                    {log.is_banned ? (
+                      <span className="flex items-center gap-1.5">
+                        <span className="text-red-600 dark:text-red-400 font-mono text-sm">{log.client_ip}</span>
+                        <span className="px-1.5 py-0.5 bg-red-100 dark:bg-red-900/30 text-red-700 dark:text-red-400 text-xs font-medium rounded">
+                          Banned
+                        </span>
+                      </span>
+                    ) : (
+                      <span className="text-sm text-slate-600 dark:text-slate-300 font-mono">{log.client_ip || '-'}</span>
+                    )}
+                  </td>
+                  <td className="px-4 py-3 text-sm text-slate-600 dark:text-slate-300 whitespace-nowrap">
+                    {log.geo_country_code || '-'}
+                  </td>
+                  <td className="px-4 py-3 text-sm text-slate-700 dark:text-slate-300 whitespace-nowrap">
+                    {log.host || '-'}
+                  </td>
+                  <td className="px-4 py-3 text-sm text-red-700 dark:text-red-400 font-mono max-w-[250px] truncate" title={log.request_uri}>
+                    {log.request_uri || '-'}
+                  </td>
+                  <td className="px-4 py-3 whitespace-nowrap">
+                    {log.status_code && <StatusCodeBadge code={log.status_code} />}
+                  </td>
+                </tr>
+              ))}
+            </tbody>
+          </table>
+        </div>
+      )}
+
+      {/* Pagination */}
+      {data && (data.has_more || page > 1) && (
+        <div className="px-4 py-3 border-t border-slate-200 dark:border-slate-700 flex items-center justify-between bg-slate-50 dark:bg-slate-800">
+          <div className="text-sm text-slate-500 dark:text-slate-400">
+            {t('pageInfo', { page: data.page, total: data.total_pages, count: data.total })}{data.has_more && '+'}
+          </div>
+          <div className="flex gap-2">
+            <button
+              onClick={() => onPageChange((p) => Math.max(1, p - 1))}
+              disabled={page === 1}
+              className="px-3 py-1 border border-slate-300 dark:border-slate-600 rounded text-sm disabled:opacity-50 disabled:cursor-not-allowed hover:bg-slate-100 dark:hover:bg-slate-700 dark:text-slate-300"
+            >
+              {t('previous')}
+            </button>
+            <button
+              onClick={() => onPageChange((p) => p + 1)}
+              disabled={!data.has_more}
+              className="px-3 py-1 border border-slate-300 dark:border-slate-600 rounded text-sm disabled:opacity-50 disabled:cursor-not-allowed hover:bg-slate-100 dark:hover:bg-slate-700 dark:text-slate-300"
+            >
+              {t('next')}
+            </button>
+          </div>
+        </div>
+      )}
+    </div>
+  );
+}
